refactored and added mobile support
