--- frontend/src/pages/HomeNavbar.css
@@ -1,29 +1,0 @@
-:root {
-    --intense: #292f63;
-    --summer: #428dc6;
-    --just-vibe: #53b2d4;
-    --cloro: #aae5f2;
-    --smooth: #d2eff7;
-    --white: #ffffff;
-    --font-type: 'Poppins', sans-serif;
-}
-
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    font-family: var(--font-type);
-}
-
-.custom-nav-link {
-    color: var(--intense);
-    margin: 0px 5px;
-}
-
-.navbar-nav .nav-link {
-    color: var(--intense) !important;
-}
-
-.dropdown-toggle::after {
-    color: var;
-}
