+++ frontend/src/pages/profile/ChangePassword.css
@@ -1,4 +1,4 @@
-@import './Users.css';
+@import '../users/Users.css';
 
 .password-page {
     min-height: 100vh;
@@ -207,7 +207,73 @@
 }
 
 @media (max-width: 768px) {
-  .password-form-wrapper {
-    width: 90%;
-  }
+    .password-form-wrapper {
+        width: 90%;
+    }
+
+    .password-form-title {
+        font-size: 22px;
+    }
+
+    .password-requirements h3 {
+        font-size: 15px;
+    }
+
+    .password-requirements li {
+        font-size: 13px;
+    }
+
+    .form-group label {
+        font-size: 15px;
+    }
+
+    .form-group input {
+        height: 44px;
+        font-size: 15px;
+    }
+
+    .update-button {
+        height: 44px;
+        font-size: 15px;
+    }
+}
+
+@media (max-width: 480px) {
+    .password-form-wrapper {
+        width: 95%;
+        padding: 1.5rem;
+    }
+
+    .password-form-title {
+        font-size: 20px;
+        margin-bottom: 1rem;
+    }
+
+    .password-form {
+        gap: 1rem;
+    }
+
+    .password-requirements {
+        padding: 1rem;
+    }
+
+    .password-requirements h3 {
+        font-size: 14px;
+        margin-bottom: 0.5rem;
+    }
+
+    .form-group input {
+        height: 40px;
+        font-size: 14px;
+    }
+
+    .update-button {
+        height: 40px;
+        font-size: 14px;
+    }
+
+    .error-message, .success-message {
+        padding: 0.8rem;
+        font-size: 13px;
+    }
 }
