+++ frontend/src/pages/profile/UserProfile.css
@@ -1,4 +1,4 @@
-@import './Users.css';
+@import '../users/Users.css';
 
 :root {
     --intense: #292f63;
@@ -303,21 +303,95 @@
 }
 
 @media (max-width: 768px) {
-  .profile-card-wrapper {
-    width: 90%;
-  }
-  
-  .profile-info-grid {
-    grid-template-columns: 1fr;
-  }
-  
-  .profile-form-header, .profile-info-header {
-    flex-direction: column;
-    gap: 1rem;
-    align-items: flex-start;
-  }
-  
-  .edit-button, .cancel-button {
-    width: 100%;
-  }
+    .profile-card-wrapper {
+        width: 90%;
+    }
+
+    .profile-info-grid {
+        grid-template-columns: 1fr;
+    }
+
+    .profile-form-header, .profile-info-header {
+        flex-direction: column;
+        gap: 1rem;
+        align-items: flex-start;
+    }
+
+    .profile-form-header h2, .profile-info-header h2 {
+        font-size: 22px;
+    }
+
+    .edit-button, .cancel-button {
+        width: 100%;
+    }
+
+    .info-section h3 {
+        font-size: 16px;
+    }
+
+    .info-row {
+        flex-direction: column;
+        align-items: flex-start;
+        gap: 0.3rem;
+    }
+
+    .form-group label {
+        font-size: 15px;
+    }
+
+    .form-group input {
+        height: 44px;
+        font-size: 15px;
+    }
+}
+
+@media (max-width: 480px) {
+    .profile-card-wrapper {
+        width: 95%;
+    }
+
+    .profile-form-wrapper, .info-card {
+        padding: 1.5rem;
+    }
+
+    .profile-form-header h2, .profile-info-header h2 {
+        font-size: 20px;
+    }
+
+    .profile-form {
+        gap: 1rem;
+    }
+
+    .info-section h3 {
+        font-size: 15px;
+        margin-bottom: 0.8rem;
+    }
+
+    .info-row {
+        margin-bottom: 0.7rem;
+    }
+
+    .info-label, .info-value {
+        font-size: 14px;
+    }
+
+    .action-button {
+        padding: 0.7rem 0.8rem;
+        font-size: 14px;
+    }
+
+    .form-group input {
+        height: 40px;
+        font-size: 14px;
+    }
+
+    .save-button {
+        height: 40px;
+        font-size: 14px;
+    }
+
+    .error-message, .success-message {
+        padding: 0.8rem;
+        font-size: 13px;
+    }
 }
