+++ frontend/src/pages/users/UserDetails.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {useAuth} from '../context/authContext';
+import {useAuth} from '../../context/authContext';
 
 const UserDetails = ({ 
   user,
