+++ frontend/src/pages/users/Users.css
@@ -7,7 +7,6 @@
     --text-color: #11184D;
     --footer-color: #292F63;
     --font-type: 'Poppins', sans-serif;
-;
 }
 
 * {
@@ -206,10 +205,12 @@ a:hover{
 .user{
     margin-top: 39px;
 }
+
 .user-details{
     display: flex;
     justify-content: space-between;
 }
+
 .user-title{
     font-size: 28px;
     color: var(--text-color);
@@ -311,3 +312,46 @@ a:hover{
     border-top-right-radius: 30px;
 }
 
+@media screen and (max-width: 485px) {
+    .users-list {
+        width: 90%;
+    }
+
+    .header-container {
+        flex-direction: column;
+        padding: 20px;
+    }
+
+    .header-text {
+        width: 100%;
+        margin-right: 0;
+        text-align: center;
+        margin-bottom: 20px;
+    }
+
+    .himg {
+        width: 300px;
+        height: 300px;
+    }
+
+    .user-details {
+        flex-direction: column;
+    }
+    
+    .user-title {
+        font-size: 22px;
+        width: 100%;
+        word-wrap: break-word;
+    }
+    
+    .user-description {
+        font-size: 16px;
+        width: 100%;
+        word-wrap: break-word;
+    }
+    
+    .user-header {
+        width: 100%;
+        font-size: 20px;
+    }
+}
